% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classlab}
\alias{classlab}
\alias{classlab}
\alias{classlab}
\alias{classlab}
\title{iradbpm.rda}
\format{An object of class \code{matrix} with 503 rows and 1 columns.}
\usage{
classlab

classlab

classlab

classlab
}
\description{
num 503x1

num 503x1

num 503x1

num 503x1
}
\keyword{datasets}
