% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comment}
\alias{comment}
\alias{comment}
\alias{comment}
\alias{comment}
\title{char}
\format{An object of class \code{character} of length 1.}
\usage{
comment

comment

comment

comment
}
\description{
char

char

char

char
}
\keyword{datasets}
