% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\alias{data}
\title{pollen.rda}
\format{An object of class \code{matrix} with 384 rows and 17 columns.}
\usage{
data

data
}
\description{
pollen.rda

384x17
}
\keyword{datasets}
