% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Diameter}
\alias{Diameter}
\title{playfair.rda
num 22x1}
\format{An object of class \code{matrix} with 22 rows and 1 columns.}
\usage{
Diameter
}
\description{
playfair.rda
num 22x1
}
\keyword{datasets}
