% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getCell.dim}
\alias{getCell.dim}
\title{get the dimensions of an object in a Matlab .mat file
It will return a vector of integers
Uses a data frame generated by matInfo()}
\usage{
getCell.dim(df, cellName)
}
\description{
get the dimensions of an object in a Matlab .mat file
It will return a vector of integers
Uses a data frame generated by matInfo()
}
