% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labs}
\alias{labs}
\alias{labs}
\title{hamster.rda}
\format{An object of class \code{matrix} with 22 rows and 1 columns.}
\usage{
labs

labs
}
\description{
hamster.rda

char 22x1
}
\keyword{datasets}
