% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LivestockName}
\alias{LivestockName}
\title{char 1 x 5}
\format{An object of class \code{matrix} with 5 rows and 1 columns.}
\usage{
LivestockName
}
\description{
char 1 x 5
}
\keyword{datasets}
