% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LivestockType}
\alias{LivestockType}
\title{num 1 x 130}
\format{An object of class \code{matrix} with 130 rows and 1 columns.}
\usage{
LivestockType
}
\description{
num 1 x 130
}
\keyword{datasets}
