% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab.R
\name{m.repmat}
\alias{m.repmat}
\title{Replicates the repmat function in Matlab}
\usage{
m.repmat(M, n, m)
}
\arguments{
\item{M}{a matrix}

\item{n}{number of rows}

\item{m}{number of columns}
}
\description{
Replicates the repmat function in Matlab
}
