% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab.R
\name{m.zeros}
\alias{m.zeros}
\title{Replicate the Matlab zeros function}
\usage{
m.zeros(m, n)
}
\arguments{
\item{m}{number of rows}

\item{n}{number of columns}
}
\description{
Replicate the Matlab zeros function
}
