% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R, R/functions.R
\name{matInfo}
\alias{matInfo}
\alias{matInfo}
\title{get information of .mat file objects
as a data frame}
\usage{
matInfo(matfile)

matInfo(matfile)
}
\description{
get information of .mat file objects
as a data frame

get information of .mat file objects
as a data frame
}
