% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ochiaibpm}
\alias{ochiaibpm}
\title{matrix 503x503}
\format{An object of class \code{matrix} with 503 rows and 503 columns.}
\usage{
ochiaibpm
}
\description{
matrix 503x503
}
\keyword{datasets}
