% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SolarRadiation}
\alias{SolarRadiation}
\title{num, 1x111}
\format{An object of class \code{matrix} with 111 rows and 1 columns.}
\usage{
SolarRadiation
}
\description{
num, 1x111
}
\keyword{datasets}
