% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{split.matrix}
\alias{split.matrix}
\title{split a 1-D matrix in a 2-D matrix by means of the # columns}
\usage{
\method{split}{matrix}(mat, cols)
}
\description{
split a 1-D matrix in a 2-D matrix by means of the # columns
}
