% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X}
\alias{X}
\alias{X}
\alias{X}
\title{data.rda}
\format{An object of class \code{matrix} with 3 rows and 800 columns.}
\usage{
X

X

X
}
\description{
data.rda

num 6x5

3x2000
}
\keyword{datasets}
